Implement AbstractFactoryProduct
